requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.11.3
python-multipart==0.0.9
httpx==0.27.0
PyJWT==2.8.0
bcrypt==4.1.3
boto3==1.35.24
stripe==9.9.0

